Add Initial Lexer For Zoid Language
Created a lexer for part of a simple language with accompanying tests to
verify the correct functioning of the lexer.

Currently it supports numeric literals (floats and integers) along with
basic operators, keywords, and identifiers. The supported keywords are
for function declaration (`fn`), `return` statements, if statements (`if`
and `else`), loops (`while`), and variable declaration (`let`).

Still needed to be added to the lexer for basic functioning are brackets
and braces (`()`, `{}`, `[]`) along with support for adding types to the
variables. The types will use `:` as the delimiter. An additional token
of the semicolon (`;`) will be used to help differentiate statements.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>